src/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication (future)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client for Supabase
httpx==0.26.0

# Utilities
python-dateutil==2.8.2